sytstemID_requirements.txt
Sphinx==1.8.1
numpydoc==0.9.2
sphinx_rtd_theme==0.4.3
sphinx-autoapi
scikit-learn 
jupyter 
h5py

